SDIT-4164: ✨ Add in offender identifier mappings (#787)

+++ src/main/kotlin/uk/gov/justice/digital/hmpps/nomismappingservice/coreperson/OffenderIdentifierMappingRepository.kt
@@ -5,8 +5,8 @@ import org.springframework.stereotype.Repository
 
 @Repository
 interface OffenderIdentifierMappingRepository : CoroutineCrudRepository<OffenderIdentifierMapping, String> {
-  suspend fun findOneByNomisOffenderIdAndNomisIdentifierSequence(nomisOffenderId: Long, nomisIdentifierSequence: Long): OffenderIdentifierMapping?
+  suspend fun findOneByNomisOffenderIdAndNomisIdentifierSequence(nomisOffenderId: Long, nomisIdentifierSequence: Int): OffenderIdentifierMapping?
   suspend fun findOneByCprId(cprId: String): OffenderIdentifierMapping?
-  suspend fun deleteByNomisOffenderIdAndNomisIdentifierSequence(nomisOffenderId: Long, nomisIdentifierSequence: Long)
+  suspend fun deleteByNomisOffenderIdAndNomisIdentifierSequence(nomisOffenderId: Long, nomisIdentifierSequence: Int)
   suspend fun deleteAllByNomisPrisonNumber(nomisPrisonNumber: String)
 }
